+++ test/test-service-lock.sh
@@ -260,7 +260,7 @@ new "Wait until timeout"
 sleep 10
 
 new "Verify action timeout"
-ret=$($clixon_cli -1 -f $CFG -E $CFD show transaction)
+ret=$($clixon_cli -1 -f $CFG -E $CFD show transaction detail)
 match=$(echo "$ret" | grep --null -Eo "<reason>Timeout waiting for service daemon</reason>") || true
 if [ -z "$match" ]; then
     err "<Timeout>" "$ret"